moved items from controllers folder to utils
